MEJOR UNLINK Y SEGUR Y G ADM PARA ESTADOS ADJ ALEX

+++ addons/dont_unlink_moves/models/account_move.py
@@ -1,12 +1,21 @@
-from odoo import _, api, fields, models
+from odoo import _, models
 from odoo.exceptions import UserError
+
 
 class AccountMove(models.Model):
     _inherit = 'account.move'
 
     def unlink(self):
-        # Levantar error para evitar cualquier eliminación
-        raise UserError(_('No está permitido eliminar asientos contables. Esta Accion Sera Reportada al administrador.'))
-        
-        # El return nunca se ejecutará por el raise, pero se mantiene por convención
-        return super(AccountMove, self).unlink()
+
+        # Permitir eliminaciones internas de Odoo
+        # durante procesos como ROMPER CONCILIACIÓN
+        if self.env.context.get('force_delete'):
+            return super(AccountMove, self).unlink()
+
+        # Bloquear eliminación manual de asientos
+        raise UserError(
+            _(
+                'No está permitido eliminar asientos contables. '
+                'Esta Acción Será Reportada al administrador.'
+            )
+        )
